+++ packages/client/src/common/utils/project.ts
@@ -1,0 +1,5 @@
+import { BaseProject } from '$stores/project';
+
+export function filterDraft(projects: BaseProject[]) {
+  return projects.filter(project => project.draft && project.changed);
+}