+++ packages/client/src/components/projectMenu/index.tsx
@@ -1,5 +1,5 @@
 import { Project } from '@metastable/types';
-import React from 'react';
+import React, { useCallback } from 'react';
 import { ContextMenuDivider, ContextMenuItem } from 'use-context-menu';
 
 import { ProjectDelete } from '$modals/project/delete';
@@ -7,15 +7,27 @@ import { ProjectDuplicate } from '$modals/project/duplicate';
 import { ProjectRename } from '$modals/project/rename';
 import { mainStore } from '$stores/MainStore';
 import { modalStore } from '$stores/ModalStore';
+import { BaseProject } from '$stores/project';
 
 export interface ProjectMenuProps {
-  project: Project;
+  projectData?: Project;
+  project?: BaseProject;
+  isTab?: boolean;
 }
 
-export const ProjectMenu: React.FC<ProjectMenuProps> = ({ project }) => {
-  const getProjectObj = () => {
-    return mainStore.projects.get(project.id);
-  };
+export const ProjectMenu: React.FC<ProjectMenuProps> = ({
+  projectData,
+  project,
+  isTab,
+}) => {
+  const getProjectObj = useCallback(
+    () => project ?? mainStore.projects.get(projectData!.id),
+    [project, projectData],
+  );
+
+  if (!project && !projectData) {
+    return null;
+  }
 
   return (
     <>
@@ -33,7 +45,6 @@ export const ProjectMenu: React.FC<ProjectMenuProps> = ({ project }) => {
       >
         Rename
       </ContextMenuItem>
-      <ContextMenuDivider />
       <ContextMenuItem
         onSelect={async () => {
           modalStore.show(<ProjectDelete project={await getProjectObj()} />);
@@ -41,6 +52,17 @@ export const ProjectMenu: React.FC<ProjectMenuProps> = ({ project }) => {
       >
         Delete
       </ContextMenuItem>
+      {isTab && !!project && (
+        <>
+          <ContextMenuDivider />
+          <ContextMenuItem onSelect={() => project.close()}>
+            Close
+          </ContextMenuItem>
+          <ContextMenuItem onSelect={() => project.closeOther()}>
+            Close other tabs
+          </ContextMenuItem>
+        </>
+      )}
     </>
   );
 };