+++ packages/client/src/modals/project/unsaved/index.tsx
@@ -3,56 +3,59 @@ import React from 'react';
 
 import { Button } from '$components/button';
 import { Modal, ModalActions, useModal } from '$components/modal';
-import { mainStore } from '$stores/MainStore';
+import { BaseProject } from '$stores/project';
+import { filterDraft } from '$utils/project';
 
-export const ProjectUnsaved: React.FC = observer(() => {
-  const { close } = useModal();
-  const draftProjects = mainStore.projects.draft;
+interface ProjectUnsavedProps {
+  projects: BaseProject[];
+  onClose?: () => void;
+}
 
-  return (
-    <Modal title="Unsaved projects" size="small">
-      <div>The following projects will be lost after closing:</div>
-      <ul>
-        {draftProjects.map(project => (
-          <li key={project.id}>{project.name}</li>
-        ))}
-      </ul>
-      <div>Would you like to save them?</div>
-      <ModalActions>
-        <Button variant="secondary" onClick={() => close()}>
-          Cancel
-        </Button>
-        <Button
-          variant="danger"
-          onClick={async () => {
-            for (const project of draftProjects) {
-              project.close(true);
-            }
+export const ProjectUnsaved: React.FC<ProjectUnsavedProps> = observer(
+  ({ projects, onClose }) => {
+    const { close } = useModal();
 
-            await Promise.all(draftProjects.map(project => project.delete()));
-            close();
-            mainStore.exit(true);
-          }}
-        >
-          Delete
-        </Button>
-        <Button
-          variant="primary"
-          onClick={async () => {
-            for (const project of draftProjects) {
-              project.close(true);
-            }
+    return (
+      <Modal title="Unsaved projects" size="small">
+        <div>The following projects will be lost after closing:</div>
+        <ul>
+          {filterDraft(projects).map(project => (
+            <li key={project.id}>{project.name}</li>
+          ))}
+        </ul>
+        <div>Would you like to save them?</div>
+        <ModalActions>
+          <Button variant="secondary" onClick={() => close()}>
+            Cancel
+          </Button>
+          <Button
+            variant="danger"
+            onClick={async () => {
+              await Promise.all(projects.map(project => project.delete()));
+              close();
+              onClose?.();
+            }}
+          >
+            Delete
+          </Button>
+          <Button
+            variant="primary"
+            onClick={async () => {
+              await Promise.all(
+                projects.map(project => project.save(undefined, false)),
+              );
+              for (const project of projects) {
+                project.close(true);
+              }
 
-            await Promise.all(
-              draftProjects.map(project => project.save(undefined, false)),
-            );
-            close();
-            mainStore.exit(true);
-          }}
-        >
-          Save
-        </Button>
-      </ModalActions>
-    </Modal>
-  );
-});
+              close();
+              onClose?.();
+            }}
+          >
+            Save
+          </Button>
+        </ModalActions>
+      </Modal>
+    );
+  },
+);