+++ packages/client/src/common/api/index.ts
@@ -1,21 +1,19 @@
-import { createWSClient, httpLink, splitLink, wsLink } from '@trpc/client';
+import {
+  createWSClient,
+  httpLink,
+  splitLink,
+  TRPCLink,
+  TRPCWebSocketClient,
+  wsLink,
+} from '@trpc/client';
 import { createTRPCReact } from '@trpc/react-query';
 import { toJS } from 'mobx';
 import { ipcLink } from 'trpc-electron/renderer';
 
 import { IS_ELECTRON } from '$utils/config';
+import { BasicEventEmitter } from '$utils/events';
 import { getUrl } from '$utils/url';
 import type { Router } from '@metastable/metastable';
-
-declare global {
-  interface Window {
-    wsIsOpen: boolean;
-    wsOnOpen?: () => void;
-    wsOnClose?: () => void;
-  }
-}
-
-window.wsIsOpen = false;
 
 const transformer = {
   serialize(data: any) {
@@ -30,31 +28,65 @@ const transformer = {
   },
 };
 
-const link = IS_ELECTRON
-  ? ipcLink({ transformer })
-  : splitLink({
-      condition: op => op.type === 'subscription',
-      true: wsLink({
-        client: createWSClient({
-          url: getUrl('/trpc', 'ws'),
-          onOpen: () => {
-            window.wsIsOpen = true;
-            window.wsOnOpen?.();
-          },
-          onClose: () => {
-            window.wsIsOpen = false;
-            window.wsOnClose?.();
-          },
+class LinkManager extends BasicEventEmitter<{
+  connectionStateChange: (isConnected: boolean) => void;
+}> {
+  link: TRPCLink<Router>;
+  wsClient?: TRPCWebSocketClient;
+  tokenCallback?: () => Promise<string | undefined> | string | undefined;
+  private _connected = true;
+
+  constructor() {
+    super();
+
+    if (IS_ELECTRON) {
+      this.link = ipcLink({ transformer });
+    } else {
+      const getHeaders = async () => ({
+        authorization: await this.tokenCallback?.(),
+      });
+
+      this._connected = false;
+      this.wsClient = createWSClient({
+        url: getUrl('/trpc', 'ws'),
+        onOpen: () => {
+          this._connected = true;
+          this.emit('connectionStateChange', true);
+        },
+        onClose: () => {
+          this._connected = false;
+          this.emit('connectionStateChange', false);
+        },
+        connectionParams: getHeaders,
+      });
+
+      this.link = splitLink({
+        condition: op => op.type === 'subscription',
+        true: wsLink({
+          client: this.wsClient,
+          transformer,
         }),
-        transformer,
-      }),
-      false: httpLink({
-        url: getUrl('/trpc'),
-        transformer,
-      }),
-    });
+        false: httpLink({
+          url: getUrl('/trpc'),
+          transformer,
+          headers: getHeaders,
+        }),
+      });
+    }
+  }
+
+  get isConnected() {
+    return this._connected;
+  }
+
+  reconnect() {
+    this.wsClient?.reconnect(null);
+  }
+}
+
+export const linkManager = new LinkManager();
 
 export const TRPC = createTRPCReact<Router>();
 export const API = TRPC.createClient({
-  links: [link],
+  links: [linkManager.link],
 });