+++ packages/client/src/views/system/Main.tsx
@@ -5,6 +5,7 @@ import { LoadingOverlay } from '$components/loadingOverlay';
 import { mainStore } from '$stores/MainStore';
 import { uiStore } from '$stores/UIStore';
 import { IS_ELECTRON } from '$utils/config';
+import { LogIn } from './LogIn';
 import styles from './Main.module.scss';
 import { Status } from './Status';
 import { SystemMonitor } from './systemMonitor';
@@ -13,6 +14,17 @@ import { TitleBar } from './TitleBar';
 
 export const Main: React.FC<React.PropsWithChildren> = observer(
   ({ children }) => {
+    if (mainStore.authorizationRequired) {
+      return (
+        <div className={styles.main}>
+          <TitleBar />
+          <div className={styles.wrapper}>
+            <LogIn />
+          </div>
+        </div>
+      );
+    }
+
     const isReady = mainStore.isConfigured;
 
     if (!IS_ELECTRON && !mainStore.ready) {