+++ CHANGELOG.md
@@ -9,6 +9,7 @@
   - `Ctrl+Left` - switches to the previous project in the project list
   - `Ctrl+Right` - switches to the next project in the project list
 - Added support for using Shift to skip the save dialog while closing draft projects.
+- Added an option to close other tabs in the context menu.
 
 ## Bug fixes